requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.2
passlib[argon2]>=1.7.4
pyjwt>=2.9.0
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.0.0
slowapi>=0.1.9
uuid>=1.30
pydantic-settings>=2.0.0
pydantic>=1.10
python-jose>=3.3.0
email-validator>=2.0.0
python-multipart>=0.0.9
watchfiles